Move transitive dependency predicates to AbstractNotationDependency
