+++ src/test/groovy/com/github/blindpirate/gogradle/core/dependency/AbstractNotationDependencyTest.groovy
@@ -1,13 +1,14 @@
 package com.github.blindpirate.gogradle.core.dependency
 
 import com.github.blindpirate.gogradle.core.GolangPackage
-import com.github.blindpirate.gogradle.core.dependency.produce.strategy.DependencyProduceStrategy
 import com.github.blindpirate.gogradle.core.dependency.resolve.DependencyResolver
 import com.github.blindpirate.gogradle.util.DependencyUtils
 import com.github.blindpirate.gogradle.util.MockUtils
+import com.github.blindpirate.gogradle.util.ReflectionUtils
 import org.junit.Before
 import org.junit.Test
 
+import static com.github.blindpirate.gogradle.core.dependency.AbstractNotationDependency.PropertiesExclusionPredicate
 import static org.mockito.Mockito.*
 
 class AbstractNotationDependencyTest {
@@ -15,7 +16,7 @@ class AbstractNotationDependencyTest {
 
     @Before
     void setUp() {
-        DependencyUtils.setExclusionSpecs(depedency, [] as Set)
+        ReflectionUtils.setField(depedency, 'transitiveDepExclusions', [] as Set)
     }
 
     @Test
@@ -24,7 +25,7 @@ class AbstractNotationDependencyTest {
         depedency.setTransitive(false)
         // then
         // exclude any transitive dependencis
-        assert depedency.getTransitiveDepExclusions().first().isSatisfiedBy(null)
+        assert depedency.getTransitiveDepExclusions().first().test(null)
     }
 
     @Test
@@ -35,7 +36,7 @@ class AbstractNotationDependencyTest {
         // when
         depedency.exclude([name: 'a'])
         // then
-        assert depedency.getTransitiveDepExclusions().first().isSatisfiedBy(dependency)
+        assert depedency.getTransitiveDepExclusions().first().test(dependency)
     }
 
     @Test
@@ -48,9 +49,20 @@ class AbstractNotationDependencyTest {
 
     static class NotationDependencyForTest extends AbstractNotationDependency {
         private static final long serialVersionUID = 1
+
         @Override
         protected Class<? extends DependencyResolver> getResolverClass() {
             return null
         }
+    }
+
+    @Test
+    void 'multiple PropertiesExclusionSpec should be compared properly'() {
+        PropertiesExclusionPredicate spec1 = PropertiesExclusionPredicate.of([name: 'name'])
+        PropertiesExclusionPredicate spec2 = PropertiesExclusionPredicate.of([name: 'name'] as TreeMap)
+        assert spec1.equals(spec2)
+        assert spec1.equals(spec1)
+        assert !spec1.equals(null)
+        assert spec1.hashCode() == spec2.hashCode()
     }
 }
